+++ valhalla/loki/worker.h
@@ -73,7 +73,7 @@ protected:
   std::unordered_map<std::string, size_t> max_locations;
   std::unordered_map<std::string, float> max_distance;
   std::unordered_map<std::string, float> max_matrix_distance;
-  float max_timedep_dist_matrix;
+  size_t max_timedep_dist_matrix;
   std::unordered_map<std::string, float> max_matrix_locations;
   size_t max_exclude_locations;
   float max_exclude_polygons_length;
